+++ src/main/java/org/gservlet/FileWatcher.java
@@ -19,32 +19,26 @@
 
 package org.gservlet;
 
-import static java.nio.file.StandardWatchEventKinds.ENTRY_CREATE;
-import static java.nio.file.StandardWatchEventKinds.ENTRY_DELETE;
-import static java.nio.file.StandardWatchEventKinds.ENTRY_MODIFY;
 import static org.gservlet.Constants.RELOAD;
 import java.io.File;
-import java.io.IOException;
-import java.nio.file.ClosedWatchServiceException;
-import java.nio.file.FileSystems;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.WatchEvent;
-import java.nio.file.WatchKey;
-import java.nio.file.WatchService;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
+import org.apache.commons.io.monitor.FileAlterationListener;
+import org.apache.commons.io.monitor.FileAlterationListenerAdaptor;
+import org.apache.commons.io.monitor.FileAlterationMonitor;
+import org.apache.commons.io.monitor.FileAlterationObserver;
 
 /**
  * 
- * Checks a folder for file changes and notifies the file
- * listeners accordingly
+ * Checks a folder for file changes and notifies the file listeners accordingly
  * 
  * @author Mamadou Lamine Ba
  * 
  */
-public class FileWatcher implements Runnable {
+public class FileWatcher {
 
 	/**
 	 * The list of file listeners
@@ -56,19 +50,25 @@ public class FileWatcher implements Runnable {
 	protected final File folder;
 
 	/**
-	 * The list of watch services
+	 * The list of FileWatcher instances
 	 */
-	protected static final List<WatchService> watchServices = new ArrayList<>();
-	
+	protected static final List<FileWatcher> instances = new ArrayList<>();
+
+	/**
+	 * The FileAlterationMonitor instance
+	 */
+	protected FileAlterationMonitor monitor = new FileAlterationMonitor(1000);
+
 	/**
 	 * 
 	 * Constructs a FileWatcher for the given folder
 	 * 
-	 * @param folder the folder object
+	 * @param folder the folder to watch
 	 * 
 	 */
 	public FileWatcher(File folder) {
 		this.folder = folder;
+		instances.add(this);
 	}
 
 	/**
@@ -79,77 +79,55 @@ public class FileWatcher implements Runnable {
 	public void watch() {
 		boolean reload = Boolean.parseBoolean(System.getenv(RELOAD));
 		if (folder.exists() && reload) {
-			Thread thread = new Thread(this);
-			thread.setDaemon(true);
-			thread.start();
+			try {
+				FileAlterationObserver observer = new FileAlterationObserver(folder.getAbsolutePath());
+				FileAlterationListener listener = new FileAlterationListenerAdaptor() {
+
+					final Set<File> files = new HashSet<>();
+
+					@Override
+					public void onFileCreate(File file) {
+						listeners.forEach(listener -> listener.onCreated(new FileEvent(file)));
+					}
+
+					@Override
+					public void onFileDelete(File file) {
+						listeners.forEach(listener -> listener.onDeleted(new FileEvent(file)));
+					}
+
+					@Override
+					public void onFileChange(File file) {
+						if (!files.contains(file)) {
+							listeners.forEach(listener -> listener.onModified(new FileEvent(file)));
+							files.add(file);
+						}
+					}
+
+					@Override
+					public void onStop(FileAlterationObserver observer) {
+						files.clear();
+					}
+
+				};
+				observer.addListener(listener);
+				monitor.addObserver(observer);
+				monitor.start();
+			} catch (Exception e) {
+				// the exception is ignored
+			}
 		}
 	}
-
+	
 	/**
 	 * 
-	 * Used by a daemon thread to start the watch process
+	 * Stops the watch process
 	 * 
 	 */
-	@Override
-	public void run() {
-		try (WatchService watchService = FileSystems.getDefault().newWatchService()) {
-			Path path = Paths.get(folder.getAbsolutePath());
-			path.register(watchService, ENTRY_CREATE, ENTRY_MODIFY, ENTRY_DELETE);
-			watchServices.add(watchService);
-			boolean poll = true;
-			while (poll) {
-				poll = pollEvents(watchService);
-			}
-		} catch (IOException | InterruptedException | ClosedWatchServiceException e) {
-			Thread.currentThread().interrupt();
-		}
-	}
-
-	/**
-	 * 
-	 * Polls for file events
-	 * 
-	 * @param watchService the watch service
-	 * @return the reset flag
-	 * @throws InterruptedException the InterruptedException 
-	 * 
-	 */
-	private boolean pollEvents(WatchService watchService) throws InterruptedException {
-		WatchKey key = watchService.take();
-		Path path = (Path) key.watchable();
-		for (WatchEvent<?> event : key.pollEvents()) {
-			notifyFileListeners(event.kind(), path.resolve((Path) event.context()).toFile());
-		}
-		return key.reset();
-	}
-
-	/**
-	 * 
-	 * Notifies the file listeners of a file event
-	 * 
-	 * @param kind the watch event kind
-	 * @param file the file upon which the event occurred upon
-	 * 
-	 */
-	private void notifyFileListeners(WatchEvent.Kind<?> kind, File file) {
-		FileEvent event = new FileEvent(file);
-		if (kind == ENTRY_CREATE) {
-			for (FileListener listener : listeners) {
-				listener.onCreated(event);
-			}
-		} 
-		else if (kind == ENTRY_MODIFY) {
-			for (FileListener listener : listeners) {
-				listener.onModified(event);
-			}
-		} 
-		else if (kind == ENTRY_DELETE) {
-			for (FileListener listener : listeners) {
-				listener.onDeleted(event);
-			}
-		}
-		if ((kind == ENTRY_CREATE || kind == ENTRY_MODIFY) && file.isDirectory()) {
-			new FileWatcher(file).addFileListeners(listeners).watch();
+	public void stop() {
+		try {
+			monitor.stop();
+		} catch (Exception e) {
+			// the exception is ignored
 		}
 	}
 
@@ -205,13 +183,13 @@ public class FileWatcher implements Runnable {
 
 	/**
 	 * 
-	 * Returns an unmodifiable list of the watch services
+	 * Returns an unmodifiable list of the FileWatcher instances
 	 * 
-	 * @return an unmodifiable list of the watch services
+	 * @return an unmodifiable list of the FileWatcher instances
 	 * 
 	 */
-	public static List<WatchService> getWatchServices() {
-		return Collections.unmodifiableList(watchServices);
+	public static List<FileWatcher> getInstances() {
+		return Collections.unmodifiableList(instances);
 	}
-	
+
 }